+++ compatibility-template/src/main/java/com/yanny/ali/plugin/template/Plugin.java
@@ -1,0 +1,12 @@
+package com.yanny.ali.plugin.template;
+
+import com.yanny.ali.api.AliEntrypoint;
+import com.yanny.ali.api.IPlugin;
+import com.yanny.ali.api.IServerRegistry;
+
+@AliEntrypoint
+public class Plugin implements IPlugin {
+    @Override
+    public void registerServer(IServerRegistry registry) {
+    }
+}